glynac-ems/requirements.txt
Django==4.2.15
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
django-environ==0.11.2
psycopg[binary]==3.2.1
django-filter==24.2
Faker==26.0.0
